Keep SelectionInfoField redux connected.
