+++ ui/appui-react/src/appui-react/statusfields/SelectionCount.tsx
@@ -9,9 +9,10 @@
 import "./SelectionCount.scss";
 import classnames from "classnames";
 import * as React from "react";
+import { FooterIndicator } from "@itwin/appui-layout-react";
+import type { IModelConnection } from "@itwin/core-frontend";
 import type { CommonProps } from "@itwin/core-react";
 import { Icon } from "@itwin/core-react";
-import { FooterIndicator } from "@itwin/appui-layout-react";
 import { SvgCursor } from "@itwin/itwinui-icons-react";
 
 /** Properties for the [[SelectionCountField]] component.
@@ -22,6 +23,7 @@ export interface SelectionCountFieldProps extends CommonProps {
 }
 
 /** Status field component used to display the number of selected items.
+ * @note Use [[useSelectionSetSize]] hook to get the selection count.
  * @beta
  */
 export function SelectionCountField(props: SelectionCountFieldProps) {
@@ -38,4 +40,34 @@ export function SelectionCountField(props: SelectionCountFieldProps) {
       {props.count}
     </FooterIndicator>
   );
+}
+
+/** Arguments for [[useSelectionSetSize]] hook.
+ * @beta
+ */
+export interface UseSelectionSetSizeArgs {
+  iModel: IModelConnection | undefined;
+}
+
+/** React hook that returns element count of a selection set.
+ * @beta
+ */
+export function useSelectionSetSize(args: UseSelectionSetSizeArgs): number {
+  const [size, setSize] = React.useState(0);
+  const { iModel } = args;
+  React.useEffect(() => {
+    if (!iModel) {
+      setSize(0);
+      return;
+    }
+    setSize(iModel.selectionSet.size);
+  }, [iModel]);
+  React.useEffect(() => {
+    if (!iModel)
+      return;
+    return iModel.selectionSet.onChanged.addListener((ev) => {
+      setSize(ev.set.size);
+    });
+  }, [iModel]);
+  return size;
 }
